Add definition and test file for StuffIt X archives.
There is an existing definition for the older StuffIt and StuffIt 5.5 formats.
This newer format was added in StuffIt 7 in 2002 with separate Media-Type.

From `StuffIt Deluxe 7.0.app/Contents/Info.plist` on my PowerBook 2400c:

```
<dict>
  <key>CFBundleTypeName</key>
  <string>StuffIt X Archive</string>
  <key>CFBundleTypeRole</key>
  <string>Editor</string>
  <key>CFBundleTypeIconFile</key>
  <string>159</string>
  <key>CFBundleTypeOSTypes</key>
  <array>
    <string>SITX</string>
  </array>
  <key>CFBundleTypeExtensions</key>
  <array>
    <string>sitx</string>
  </array>
  <key>CFBundleTypeMIMETypes</key>
  <array>
    <string>application/x-stuffitx</string>
    <string>application/x-sitx</string>
  </array>
</dict>
```

My included `test.sitx` was created with the same version of the software
on the same machine and is also added to `tests/mime-detection/list`:

```
[okeeblow@emi#shared-mime-info] xxd -l 16 tests/mime-detection/test.sitx
00000000: 5374 7566 6649 7421 a10a 4a05 9bdd 4215  StuffIt!..J...B.
```

Apache HTTPd is aware of this Media-Type:
https://bz.apache.org/bugzilla/show_bug.cgi?id=38009

PRONOM format definition:
https://www.nationalarchives.gov.uk/PRONOM/Format/proFormatSearch.aspx?status=detailReport&id=1147