Add DOS/Windows batch file type
Files of type application/x-bat are similar to files of type
application/x-ms-dos-executable in that they can be run with Wine or
DOSBox, but are different in that they are text, not binary.

application/x-bat is not an official IANA media type, but is used by
Apache Tika, which also accepts application/bat as an alias.

https://ciam-tika.straumann.com/mime-types/application/x-bat