Give higher priority to the more specific image/apng magic
Otherwise it's unclear which of image/png and image/apng should win.
The spec doesn't say that one should check if one match is a subclass
of another, for magic matching, it says priority is used for this:

"Low numbers should be used for more generic types (such as 'gzip
compressed data') and higher values for specific subtypes (such as a
word processor format that happens to use gzip to compress the file)."

Found when porting QtCore to shared-mime-info-2.3.