Add application/x-msdownload and subtypes
application/x-msdownload is what Microsoft Windows uses for EXE and DLL
files, so that type is commonly found on the WWW and in other contexts.
It is an umbrella type for anything with a .exe or .dll file extension,
among others.

application/x-dosexec is what the `file` utility uses for DOS
executables specifically. According to the File Format Problem wiki:
"It's not clear if there is any completely reliable way to identify a
file as strictly DOS EXE, except in the negative (i.e., it looks like
EXE, and is not a valid NE, PE, etc., file)...If the relocation table
offset is from 28 to 63...it is pretty certainly DOS EXE."

application/x-ms-ne-executable is what the `file` utility uses for
16-bit Windows libraries and executables.

application/vnd.microsoft.portable-executable is the IANA standard type
for 32-bit and 64-bit Windows libraries and executables, as well as EFI
executables.

With these changes, application/x-ms-dos-executable is considered
deprecated. For now it is an alias of application/x-msdownload, but in
the future it may change to be an alias of application/x-dosexec.

helloworld-*.exe was compiled with Open Watcom from the following C
source code:

    #include <stdio.h>

    int main(void)
    {
        puts("Hello World!");
        return 0;
    }

https://github.com/file/file/blob/171c73b67ca692550882e0b8ab3c2a1f16ab1e00/magic/Magdir/msdos
https://www.iana.org/assignments/media-types/application/vnd.microsoft.portable-executable
http://fileformats.archiveteam.org/wiki/MS-DOS_EXE
http://fileformats.archiveteam.org/wiki/New_Executable
http://fileformats.archiveteam.org/wiki/Portable_Executable