Allow cross-compiling after moving xgdmime to a meson subproject
It is not currently possible to build a meson subproject for the build
system. Since xdgmime is only used for the tests we can move this logic
into the build-tests conditional and disables building the tests when
cross-compiling. This workaround is not ideal but it's better than not
being able to cross-compile at all.

See https://github.com/mesonbuild/meson/issues/11121