Separate StarOffice 3-4 types from StarOffice 5 types
Types that start with application/vnd.stardivision are for StarOffice 5
and types that start with application/x-star are for StarOffice 4. The
StarOffice 3 formats are similar to the StarOffice 4 formats and I have
not found any distinct MIME type for them. StarOffice 2 used the same
file extensions but entirely different file formats that are not based
on OLE Compound Storage.

Also separate StarWriter master documents from ordinary StarWriter
documents because the two have distinct file formats.

https://www.openoffice.org/framework/documentation/mimetypes/mimetypes.html