Separate text/x-cython from text/x-python
Cython is a superset of Python that requires compilation; it cannot be
directly executed by the Python interpreter.

Cython files are identified by a .pyx, .pxd, or .pxi extension. I have
not seen anyone using .py2x or .py3x to indicate which Python version
the Cython module is compatible with.

https://github.com/github-linguist/linguist/blob/7ca3799b8b5f1acde1dd7a8dfb7ae849d3dfb4cd/lib/linguist/languages.yml#L1398
https://cython.readthedocs.io/en/latest/src/userguide/language_basics.html#cython-file-types