Add image/x-kiss-cel
GIMP supports this image file format. "Cel" is short for "celluloid" and
is not a typo.

The test suite already had a KiSS cel file and says that it is supposed
to be image/x-cel, but I could not find any examples of anyone using
image/x-cel in practice. On the other hand, I did find an example of
image/x-kiss-cel being used and it's a better name anyway because the
.cel file extension is also used for Autodesk Animator cel image files
and Lumena "cell outline" images.

https://gitlab.gnome.org/GNOME/gimp/-/blob/GIMP_2_10_38/plug-ins/common/file-cel.c
https://en.wikipedia.org/wiki/Cel
https://snisurset.net/code/abydos/formats.html
http://justsolve.archiveteam.org/wiki/KiSS_CEL
http://justsolve.archiveteam.org/wiki/Animator_PIC/CEL
http://justsolve.archiveteam.org/wiki/Lumena_CEL